fix(touch,gui): correct process_touch address and fix 32-bit CGFloat ABI mismatch
- Correct process_touch hook target address for armv7 32-bit: Changed from 0x003F1E28 (+380 bytes inside function) to 0x003F1CAC, identified from UIKit touch dispatch stack in .ips crash logs.

- Disable loading_messages hook on 32-bit: Address 0x002381cc is not loading_messages on armv7, preventing EXC_BAD_INSTRUCTION crash on touch dispatch.

- Fix 32-bit ABI layout for CoreGraphics & UIKit interop: Add architecture-dependent CGFloat (f32 on 32-bit, f64 on 64-bit), NSInteger, NSUInteger, fixing CGRect (16-byte on 32-bit vs corrupt 32-byte), CGPoint, CGSize, UIEdgeInsets, uiscreen_size, font sizing, and color parameters across gui.rs, language.rs, and menu.rs.

- Fix hook self-test in hook.rs using volatile indirect function pointer call to avoid compiler branch optimization past hook.

+++ src/meta/language.rs
@@ -9,7 +9,7 @@ use std::{
 };
 use strum::{EnumIter, EnumString, EnumVariantNames, IntoEnumIterator, IntoStaticStr};
 
-use super::gui::{Font, FontSet};
+use super::gui::{CGFloat, Font, FontSet};
 pub use fluent::fluent_args as msg_args;
 use objc::runtime::Object;
 
@@ -267,10 +267,10 @@ impl Language {
     /// Returns the set of fonts that should be used for this language.
     pub fn font_set(self) -> FontSet {
         // Define some standard sizes to work with.
-        const STD_TITLE: f64 = 25.0;
-        const STD_SMALL: f64 = 10.0;
-        const STD_TEXT: f64 = 15.0;
-        const STD_SUBTITLE: f64 = 17.0;
+        const STD_TITLE: CGFloat = 25.0;
+        const STD_SMALL: CGFloat = 10.0;
+        const STD_TEXT: CGFloat = 15.0;
+        const STD_SUBTITLE: CGFloat = 17.0;
 
         match self {
             // Czech and Slovak need a font that covers all of the accented characters. Arabic
@@ -317,7 +317,7 @@ impl Language {
                 text_font: Font::ChaletComprime,
                 text_size: STD_TEXT,
                 subtitle_font: Font::ChaletComprime,
-                subtitle_size: STD_SUBTITLE + 2.0,
+                subtitle_size: STD_SUBTITLE + (2.0 as CGFloat),
             },
 
             // Khmer characters always fall back to Khmer Sangam MN, because it's the only font
